+++ src/components/Assistant/Q1Assistant.tsx
@@ -1,11 +1,11 @@
 import { useContext, useEffect, useState } from "react";
-import { AuthData, Task } from "../../types";
+import { AuthData, Task } from "../../lib/types";
 import { TaskLabel } from "../TaskLabel/TaskLabel";
 import { Button } from "../ui/Button";
 import { AssistantAction } from "./assistant.lib";
 import { moveTask, TasksListAction } from "../containers/dnd.lib";
 import { useTodoTasksMutation } from "../TaskLabel/lib/tasks.fn";
-import { AuthContext } from "@/components/header/lib/authentication";
+import { AuthContext } from "@/lib/contexts";
 
 export function Q1Assistant({tasks}:{tasks: Task[] | undefined}) {
   const [message, setMessage] = useState<string[]>([]);
