+++ src/components/TaskLabel/DailyTaskLabel.tsx
@@ -1,4 +1,4 @@
-import { Task } from "../../types";
+import { Task } from "../../lib/types";
 
 export function DailyTaskLabel({task}:{task:Task}) {
