+++ src/components/TaskLabel/HabitTaskLabel.tsx
@@ -1,4 +1,4 @@
-import { Task } from "../../types";
+import { Task } from "../../lib/types";
 
 export function HabitTaskLabel({task}:{task:Task}) {
   return (
