+++ src/components/TaskLabel/lib/tasks.fn.ts
@@ -1,11 +1,12 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { AxiosRequestConfig, AxiosResponse } from "axios";
-import { AuthData, EMCategory, Task, TaskType } from "../../../types";
-import { AuthContext, getRequestSettings } from "../../header/lib/authentication";
+import { AuthData, EMCategory, Task, TaskType } from "../../../lib/types";
 import { Observable, Subject } from "rxjs";
 import { AllTaskTypesDataContract, SingleTaskDataContract, TaskListContract, TodoTaskDataContract } from "./task.datacontracts";
 import { useContext } from "react";
 import { axiosInstance } from "@/config/APIconfig";
+import { AuthContext } from "@/lib/contexts";
+import { getRequestSettings } from "@/lib/utils";
 
 const getTodos = async (authData:AuthData | undefined, type:string | undefined, id?: string):Promise<any[] | any> => {
   if(!authData) throw new Error("Habitica API token missing or broken, login to get it");
