+++ src/components/containers/SidebarTasksList.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { EMCategory, Task } from "../../types";
+import { EMCategory, Task } from "../../lib/types";
 import { TaskLabel } from "../TaskLabel/TaskLabel";
 import { useDroppable } from "@dnd-kit/core";
 import { getTasksFactory } from "../TaskLabel/lib/tasks.fn"
