+++ src/components/header/LoginForm.tsx
@@ -1,6 +1,7 @@
 import { Button } from "../ui/Button";
-import { updateCredentials } from "./lib/use-auth-data";
+import { updateCredentials } from "../../lib/hooks/use-auth-data";
 import { MouseEvent } from "react";
+import { UserCredentials } from "@/lib/types";
 
 export function LoginForm() {
   function handleLogin(e: MouseEvent<HTMLButtonElement, globalThis.MouseEvent>) {
@@ -8,7 +9,7 @@ export function LoginForm() {
     updateCredentials({
       username: (document.getElementById("email") as HTMLInputElement).value,
       password: (document.getElementById("password") as HTMLInputElement).value
-    });
+    } as UserCredentials);
   }
 
   return (
