+++ src/lib/contexts.ts
@@ -1,0 +1,4 @@
+import { AuthData } from "@/lib/types";
+import { createContext } from "react";
+
+export const AuthContext = createContext<AuthData | undefined>(undefined);
